# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: lttd0204@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/product-types/{productTypeId}:
    get:
      tags:
        - product-types
      summary: Get details product type by ID
      description: Get details product type by id of productTypes
      operationId: getProductTypesById
      parameters:
        - name: productTypeId
          in: path
          description: ID of product type that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    delete:
      tags:
        - product-types
      summary: Delete product type by ID
      description: Try delete product type by id of product types
      operationId: deleteProductTypes
      parameters:
        - name: productTypeId
          in: path
          description: ID of the product type that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - product-types
      summary: Update an existing product type
      description: Update an existing product type by Id
      operationId: updateProductType
      parameters:
        - name: productTypeId
          in: path
          description: product type id of product type updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product type in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductType"
          application/xml:
            schema:
              $ref: "#/components/schemas/ProductType"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ProductType"
        required: true

      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ProductType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/reviews/{reviewId}:
    get:
      tags:
        - reviews
      summary: Get details review by ID
      description: Get details review by id of reviews
      operationId: getReviewById
      parameters:
        - name: reviewId
          in: path
          description: ID of review that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    delete:
      tags:
        - reviews
      summary: Delete review by ID
      description: Try delete review by id of review
      operationId: deleteReview
      parameters:
        - name: reviewId
          in: path
          description: ID of the review that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - reviews
      summary: Update an existing review
      description: Update an existing review by Id
      operationId: updateReview
      parameters:
        - name: reviewId
          in: path
          description: review id of review updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent review in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
          application/xml:
            schema:
              $ref: "#/components/schemas/Review"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Review"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ProductType:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required:
        - name
        - slug
      xml:
        name: ProductType
    Review:
      type: object
      properties:
        content:
          type: string
        star:
          type: number
          minimum: 1
          maximum: 5
        product:
          type: string
          format: uuid
        user:
          type: string
          format: uuid
      required:
        - content
        - star
        - productID
        - userID
      xml:
        name: Review
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT